+++ content-script/index.ts
@@ -1,90 +1,7 @@
-type QueryParams = {
-  organization: string | null;
-  tenantName: string | null;
-  scLang: string | null;
-  scSite: string | null;
-  scItemId: string | null;
-};
+// Import the content script logic here
+import { datasourceViewer } from './datasource-viewer';
 
-const getQueryParam = (param: string): string | null => {
-  const queryString = window.location.search;
-  const urlParams = new URLSearchParams(queryString);
-  return urlParams.get(param);
-};
+// Start content script logics
+datasourceViewer();
 
-const constructExplorerUrl = (queryParams: QueryParams): string => {
-  const { organization, tenantName, scLang, scSite, scItemId } = queryParams;
-  return `https://explorer.sitecorecloud.io/columns?organization=${organization}&tenantName=${tenantName}&sc_lang=${scLang}&sc_site=${scSite}&sc_itemid=${scItemId}`;
-};
-
-const constructContentEditorUrl = (queryParams: QueryParams): string => {
-  const { organization, tenantName, scLang, scItemId } = queryParams;
-  return `https://xmc-${tenantName}.sitecorecloud.io/sitecore/shell/Applications/Content%20Editor.aspx?sc_bw=1&organization=${organization}&sc_lang=${scLang}&fo=${scItemId}`;
-};
-
-const createLinkSpan = (explorerUrl: string, contentEditorUrl: string): HTMLSpanElement => {
-  const linkSpan = document.createElement('span');
-  linkSpan.classList.add('xmcloud-extension-links');
-
-  const explorerLink = createLinkElement(explorerUrl, 'Explorer');
-  const contentEditorLink = createLinkElement(contentEditorUrl, 'Content Editor');
-
-  linkSpan.appendChild(document.createTextNode('Open in '));
-  linkSpan.appendChild(explorerLink);
-  linkSpan.appendChild(document.createTextNode(' or '));
-  linkSpan.appendChild(contentEditorLink);
-
-  return linkSpan;
-};
-
-const createLinkElement = (url: string, text: string): HTMLAnchorElement => {
-  const linkElement = document.createElement('a');
-  linkElement.classList.add('xmcloud-extension-link');
-  linkElement.href = url;
-  linkElement.innerText = text;
-  linkElement.target = '_blank';
-  return linkElement;
-};
-
-const observeDomChanges = () => {
-  const observer = new MutationObserver((mutations) => {
-    mutations.forEach((mutation) => {
-      mutation.addedNodes.forEach((addedNode) => {
-        if (addedNode instanceof HTMLElement && addedNode.closest('app-datasource-picker')) {
-          console.log('DOM changed from extension within app-datasource-picker');
-
-          const elementsWithDataItemId = addedNode.querySelectorAll('[data-itemid]') as NodeListOf<HTMLElement>;
-
-          elementsWithDataItemId.forEach((element: HTMLElement) => {
-            const itemId = element.getAttribute('data-itemid');
-            const existingLinkSpan = element.querySelector('.xmcloud-extension-links');
-
-            if (!existingLinkSpan && itemId) {
-              const queryParams: QueryParams = {
-                organization: getQueryParam('organization'),
-                tenantName: getQueryParam('tenantName'),
-                scLang: getQueryParam('sc_lang'),
-                scSite: getQueryParam('sc_site'),
-                scItemId: itemId,
-              };
-
-              const explorerUrl = constructExplorerUrl(queryParams);
-              const contentEditorUrl = constructContentEditorUrl(queryParams);
-
-              const linkSpan = createLinkSpan(explorerUrl, contentEditorUrl);
-              element.appendChild(linkSpan);
-            }
-          });
-        }
-      });
-    });
-  });
-
-  observer.observe(document.body, {
-    childList: true,
-    subtree: true,
-  });
-};
-
-// Start observing the DOM for changes
-observeDomChanges();
+// Add further content scripts here
